+++ library/QATools/QATools/PageObject/PageFactory.php
@@ -88,6 +88,13 @@ class PageFactory implements IPageFactory
 	protected $pageUrlMatcherRegistry;
 
 	/**
+	 * Selenium selector.
+	 *
+	 * @var SeleniumSelector
+	 */
+	protected $seleniumSelector;
+
+	/**
 	 * The current config.
 	 *
 	 * @var IConfig
@@ -118,7 +125,7 @@ class PageFactory implements IPageFactory
 			$container_or_config = $this->_createContainer();
 		}
 
-		$this->_setSession($session);
+		$this->_session = $session;
 		$this->config = $container_or_config['config'];
 
 		$this->_setAnnotationManager($container_or_config['annotation_manager']);
@@ -126,6 +133,7 @@ class PageFactory implements IPageFactory
 		$this->urlNormalizer = $container_or_config['url_normalizer'];
 		$this->pageLocator = $container_or_config['page_locator'];
 		$this->pageUrlMatcherRegistry = $container_or_config['page_url_matcher_registry'];
+		$this->seleniumSelector = $container_or_config['selenium_selector'];
 	}
 
 	/**
@@ -173,29 +181,9 @@ class PageFactory implements IPageFactory
 	 */
 	public function createDecorator(ISearchContext $search_context)
 	{
-		$locator_factory = new DefaultElementLocatorFactory($search_context);
+		$locator_factory = new DefaultElementLocatorFactory($search_context, $this->seleniumSelector);
 
 		return new DefaultPropertyDecorator($locator_factory, $this);
-	}
-
-	/**
-	 * Sets session.
-	 *
-	 * @param Session $session Session.
-	 *
-	 * @return self
-	 */
-	private function _setSession(Session $session)
-	{
-		$selectors_handler = $session->getSelectorsHandler();
-
-		if ( !$selectors_handler->isSelectorRegistered('se') ) {
-			$selectors_handler->registerSelector('se', new SeleniumSelector());
-		}
-
-		$this->_session = $session;
-
-		return $this;
 	}
 
 	/**
@@ -324,6 +312,19 @@ class PageFactory implements IPageFactory
 		$resolved_page_class = $this->pageLocator->resolvePage($class_name);
 
 		return new $resolved_page_class($this);
+	}
+
+	/**
+	 * Translates provided how/using combo into XPath.
+	 *
+	 * @param string $how   How class constant.
+	 * @param string $using Using value.
+	 *
+	 * @return string
+	 */
+	public function translateToXPath($how, $using)
+	{
+		return $this->seleniumSelector->translateToXPath($how, $using);
 	}
 
 }